schema: use :hash as the default table type when :key_type is specified
:array was the default table type. If :key_type is specified, user must
want to define a have key table. So :array wasn't the good default for
the case.

If user doesn't specify :key_type, :array is the default table
type. It's not changed.

It's a backward incompatible change. But nobody will not be
stumped. Because :key_type is specified but :type isn't specified case
is a bug of user's program.